requirements.txt
transformers>=4.44
datasets
accelerate
bitsandbytes
pdf2image
pillow
torch
torchvision
